feat: Add basic class implementation for Car
This commit adds a new file `7_basic_class.ts` which contains the implementation of a basic class `Car`. The class has a property `make` of type string and a method `drive()` that prints a message to the console. The purpose of this commit is to provide a basic class structure for a car and demonstrate its usage by creating an instance of the `Car` class and calling the `drive()` method.