feat: Add constructor to Car class
This commit adds a constructor to the `Car` class in the `8_class_with_constructor.ts` file. The constructor initializes the `make` property of the `Car` class. It also creates an instance of the `Car` class with the make set to 'Toyota' and calls the `drive()` method to print "Driving a Toyota" to the console.