feat: Make 'make' property private in Car class
This commit modifies the `Car` class in the `9_class_with_private_attributes.ts` file. The `make` property is changed to private to encapsulate it within the class. A new method `getMake()` is added to retrieve the value of the `make` property. This change improves the data encapsulation and access control in the `Car` class.